Created CRUD for entity Organization
Motivation:
The API should provide a functionality to create and modify organizations.

Modifications:
Created the organization and corresponding image model. Added the controller class and the repository interface for the organization.

Result:
Now it's possible to create, read (single or multiple (if necessary with offset and limit)), update and delete organizations via REST API calls.