Created CRUD for Service and edit ServiceProvider
Motivation:
The API should provide a functionality to create and modify a Service.

Modifications:
Extended the Service model. Created enum for LifecycleStatus and Authentication. Added the controller class and the repository interface. Changed the ServiceProvider, to get the corresponding list of Services.

Result:
Now it's possible to create, read (single or multiple (+PagingAndSorting)), update and delete Services via REST API calls. You can get a list of services for each ServiceProvider.