ignore Mac generated file and fix auth url in open-api config
Motivation:

Mac generated files can be added to this project accidentally.
Since it is of no benefit to this project this should be ignore.

Commit 30e8e9a520bac4a71aa38bcb94c5ab1cbc73b399 introduce a changed
in the url of the Helmholtz AAI. Unfortunately the auth url in
open-api configuration file was not adjusted accordingly. Hence,
the login to cerebrum api documentation was broken.

Modification:

- add .DS_Store to .gitignore file
- fix authorization url inside the open-api configuration

Result:

Ignore .DS_store and change the aai url inside the open-api
configuration.

Target: master
Acked-by: Carsten Heidmann
Review: https://gitlab.hzdr.de/hifis-technical-platform/helmholtz-cerebrum/-/merge_requests/18