Motivation: The API should provide a functionality to create and modify users.
Modifications:
Modify repository for PagingAndSorting
Extend the controller by common CRUD methods and use the existing implementations of the interface.
Use the cerebrum error handling for the user.

Result:
Now it should be possible to create, read (single or multiple (+PagingAndSorting)), update and delete users via REST API calls. The implementation is currently not tested, because of an authorization problem.